Extracted github functions for shareability
